Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/fetch/sec-metadata/sharedworker.tentative.https.sub-expected.txt

CONSOLE MESSAGE: ReferenceError: Can't find variable: SharedWorker

FAIL Untitled ReferenceError: Can't find variable: SharedWorker
